// src/app/pet/models/pet.ts
export interface Pet {
  id: number;
  name: string;
  height: string;
  mass: string;
  hair_color: string;
  skin_color: string;
  eye_color: string;
  birth_year: string;
  gender: string;
  numberOfPats: number;
  dressed: boolean;
  numberOfFeeds: number;
  numberOfWalks: number;
  washCount: number;
}